Updated table colors and table variables.
Updated table colors and color intervals in skin.conf. There is a bug in historygenerator.py that won't allow for float values for the rain scale. Can't use 0.01". This is only a problem for US units. Also, changed the location of the error pages for the CopyGenerator in Skin.conf.