Fixed bug concerning lightning data
Weewx by default sets most lightning data units to "count". Unfortunately, historygenerator.py set data with units of "count" to string and thus wouldn't properly sum the lightning strikes for each month in the table. In order to sum the lightning strikes for the table, you have to include, "units = Strikes" as well as "unit_type = meter" in the [HistoryReport] section of the skin.conf. This is just a simple hack but it works to fix this issue for lightning data. Example for Skin.conf below: 
#total monthly lightning
[[lightning_sum]]
        obs_type = lightning_strike_count
        unit_type = meter
        aggregate_type = sum
	units = Strikes
	summary_column = true
        summary_heading = Total
	colours =   "#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF"